renamed atori-tutori, and removed extra stuff.
added spawnpoint for atori tutori in horlais peak
