+++ sql/mob_pools.sql
@@ -6225,7 +6225,7 @@ INSERT INTO `mob_pools` VALUES (6229,'Il_aern_sam','Il_aern_sam',3,0x00005B05000
 INSERT INTO `mob_pools` VALUES (6230,'Il_aern_nin','Il_aern_nin',3,0x0000570500000000000000000000000000000000,13,13,9,240,100,0,0,0,1,0,0,0,0,129,5,0,7,1,0,3);
 INSERT INTO `mob_pools` VALUES (6231,'Il_aern_drg','Il_aern_drg',3,0x00005B0500000000000000000000000000000000,14,14,8,360,100,0,0,0,1,0,0,0,1,129,5,0,0,1,0,440);
 INSERT INTO `mob_pools` VALUES (6232,'Il_aern_smn','Il_aern_smn',3,0x00005C0500000000000000000000000000000000,15,15,12,360,100,0,0,0,1,0,0,0,1,129,5,0,31,1,0,439);
-INSERT INTO `mob_pools` VALUES (6233,'Atori-Tutori_???','Atori-Tutori_???',153,0x0100020500101120003000400050006000700000,2,0,1,480,110,0,0,0,0,18,0,32,7,155,0,0,0,1,0,335);
+INSERT INTO `mob_pools` VALUES (6233,'Atori-Tutori_qm','Atori-Tutori_???',153,0x0100020500101120003000400050006000700000,2,0,1,480,110,0,0,0,0,18,0,32,7,155,0,0,0,1,0,335);
 
 
 /*
